Remove aws stuff
